+++ Source/Debugger/SystemInterface.cpp
@@ -47,7 +47,7 @@ SystemInterface::~SystemInterface()
 }
 
 // Get the number of seconds elapsed since the start of the application.
-float SystemInterface::GetElapsedTime()
+double SystemInterface::GetElapsedTime()
 {
 	return application_interface->GetElapsedTime();
 }
